hewProject-main/ecsite_project/requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
defusedxml==0.7.1
django-allauth==0.58.2
django-rest-framework==0.1.0
django-widgets-improved==1.5.0
djangorestframework==3.14.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
Pillow==10.1.0
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
soupsieve==2.5
typing_extensions==4.8.0
